rewrote kernel, rewrote kernel call. added test for xor

+++ Development/GPU/FullyConnected/FullyConnectedNN.h
@@ -16,7 +16,7 @@ public:
 
     void printEmptyValues();
 
-    std::vector<size_t> getSizes();
+    std::vector<int> getSizes();
 
     std::vector<float> getAllWeights();
